[style] customizing views

+++ app/controllers/customers_controller.rb
@@ -4,7 +4,7 @@ class CustomersController < ApplicationController
   # GET /customers or /customers.json
   def index
     @customers = Customer.all
-    @pagy, @customers = pagy(@customers, items: 5)
+    @pagy, @customers = pagy(@customers, items: 10)
   end
 
   def search
@@ -43,7 +43,7 @@ class CustomersController < ApplicationController
 
     respond_to do |format|
       if @customer.save
-        format.html { redirect_to customers_path, notice: 'Customer was successfully created.' }
+        format.html { redirect_to customers_path, success: 'Customer was successfully created.' }
         format.json { render :show, status: :created, location: customers_path }
       else
         format.html { render :new, status: :unprocessable_entity }
@@ -56,7 +56,7 @@ class CustomersController < ApplicationController
   def update
     respond_to do |format|
       if @customer.update(customer_params)
-        format.html { redirect_to customer_path(@customer), notice: 'Customer was successfully updated.' }
+        format.html { redirect_to customer_path(@customer), success: 'Customer was successfully updated.' }
         format.json { render :show, status: :ok, location: @customer }
       else
         format.html { render :edit, status: :unprocessable_entity }
@@ -70,19 +70,23 @@ class CustomersController < ApplicationController
     @customer.destroy!
 
     respond_to do |format|
-      format.html { redirect_to customers_url, notice: 'Customer was successfully destroyed.' }
+      format.html { redirect_to customers_url, alert: 'Customer was successfully destroyed.' }
       format.json { head :no_content }
     end
   end
 
   def multiple_delete
-    ids = params[:customer_ids].compact
+    if params[:customer_ids].present?
+      ids = params[:customer_ids].compact
 
-    Customer.where(id: ids).destroy_all
+      Customer.where(id: ids).destroy_all
 
-    respond_to do |format|
-      format.html { redirect_to root_path, success: 'All selected Customers were successfully destroyed.' }
-      format.json { head :no_content }
+      respond_to do |format|
+        format.html { redirect_to root_path, alert: 'All selected Customers were successfully destroyed.' }
+        format.json { head :no_content }
+      end
+    else
+      flash.now[:error] = 'Please select at least one Customer.'
     end
   end
